Add user registration and vendor performance view
This commit includes an expansion of vendor and user functionalities. Firstly, the endpoint for the user registration was added with the creation of CreateUserView and its serializer UserSerializer in new files api/users/views.py and api/users/serializers.py, respectively. The new registration route was later included in the project's main urls.py file.

Secondly, a new VendorPerformanceView was implemented in api/vendors/views.py to present the performance of vendors. This endpoint uses the newly created VendorPerformanceSerializer in the api/vendors/serializers.py file. The endpoint for accessing this view has been added to the vendors' urls.py file.

Additional updates include importing Django REST Framework's authtoken to be used in the project setup and adding purchase order's signals to the api's AppConfig. The latter will be useful whenever we want to execute some code in response to signal triggers.